Refactor to include types definition by default
The current modification of mqtt_pal.h does not include the types
for the specific protocols if the custom socket is used, which is
unconvenient, as the types would be used regarless of this, and
forces to include an external header file by the 3rd party application.

This changes moves the types to the same place as the specific
protocol methods. Also, the `MQTT_PAL_ENABLE_<protocol>` definitions
are no longer needed.

In summary:
 * Several `MQTT_USE_<protocol>` definition can be used, resulting
 in definition of protocol specific types and methods.
 * In case `MQTT_USE_CUSTOM_SOCKET_HANDLE` is not defined, the part that
define `mqtt_pal_socket_handle`, `mqtt_pal_sendall` and
`mqtt_pal_recvall` into only one specific protocol is executed also,
to keep backward compatibility.
 * In case `MQTT_USE_CUSTOM_SOCKET_HANDLE` is defined, the types and
 method for the desired protocols will be available without the need
 of including an extra header.